requirements.txt
python-telegram-bot
python-decouple
sqlalchemy
psycopg2-binary